refactor: added TagData type in general monitoring types folder
