+++ src/pages/monitoring/outlets/tagsAdmin/utils/formObjectUpdate.ts
@@ -1,4 +1,4 @@
-import type { TagDataForm } from "pages/monitoring/outlets/tagsAdmin/types/tagData";
+import type { TagDataForm } from "pages/monitoring/types/tagData";
 
 export function makeInitialInfo(): TagDataForm {
   return {
